news_pipeline/requirements.txt
newspaper
python-dateutil
lxml
pymongo
redis
requests
sklearn
pika
numpy
scipy
pyyaml
simhash